move ItemConstraint to NLU module
