+++ doc/source/api/nlu.rst
@@ -9,3 +9,15 @@ NLU
    :members:
 .. automodule:: moviebot.nlu.slot_annotator
    :members:
+
+Item constraint
+-----------------
+
+.. automodule:: moviebot.nlu.annotators.item_constraint
+   :members:
+.. automodule:: moviebot.nlu.annotators.slots
+   :members:
+.. automodule:: moviebot.nlu.annotators.operator
+   :members:
+.. automodule:: moviebot.nlu.annotators.values
+   :members:
