+++ moviebot/dialogue_manager/dialogue_manager.py
@@ -8,9 +8,9 @@ __author__ = "Javeria Habib"
 from moviebot.dialogue_manager.dialogue_act import DialogueAct
 from moviebot.dialogue_manager.dialogue_policy import DialoguePolicy
 from moviebot.dialogue_manager.dialogue_state_tracker import DialogueStateTracker
-from moviebot.dialogue_manager.item_constraint import ItemConstraint
-from moviebot.dialogue_manager.operator import Operator
 from moviebot.intents.agent_intents import AgentIntents
+from moviebot.nlu.annotators.item_constraint import ItemConstraint
+from moviebot.nlu.annotators.operator import Operator
 
 
 class DialogueManager:
